relocate stream loaders to yt.loaders
